Add warning about using with `sorbet-runtime` (#194)
* Add warning about using with `sorbet-runtime`

Maybe this is overzealous, but we got really burned by it. We were using `sorbet-rails` in development only, but `sorbet-runtime` in all environments. This meant we were able to access `X::ActiveRecord_Relation` in development, and thus someone used that in a method signature. When we deployed this code it failed in production because [this railtie](https://github.com/chanzuckerberg/sorbet-rails/blob/master/lib/sorbet-rails/railtie.rb#L30) didn't run, so the constant was still private.

Adding a warning so it doesn't happen to other people.

* simpler warning