Handle NotImplementedError raises (#226)
Dearest Reviewer,

I hope all is well this November.  It as been a long month for me and I look forward to the amazing things sorbet will help me with. I have interface classes that raise NotImplementedError and the model process is dying on it. I have found locally this to be an acceptable change. I hope that you agree. 

Changes
Add NotImplementedError to the model processing exception handling for interface like classes.

Becker