Use `presence` validations to remove T.nilable on attributes and associations (#306)
This PR updates the model RBI generator for column attributes, belongs_to, and has_one to check for `presence` validations to help determine if we can avoid using `T.nilable`.

For example let's say we have a `Student` model with the following schema:
```
#  id          :bigint(8)    not null, primary key
#  first_name  :string(255)  not null
class Student
  validates :first_name, presence: true
end
```

In this case, `first_name` has a non-null DB constraint so we know the type of `first_name` is `String`. Let's pretend it there was no constraint:
```
#  id          :bigint(8)    not null, primary key
#  first_name  :string(255)
class Student
  validates :first_name, presence: true
end
```

In the model RBI that sorbet-rails generates the `first_name` would be `T.nilable(String)` which means anywhere we use first_name we need to check for nil to make the static checker happy. However, there's an `presence: true` validation on first_name which means that we can expect `first_name` to in fact not be nilable (ignoring the validation being added after some data had already been entered or columns were updated via `update_columns` to skip the validation). This PR makes use of the validation to tighten the signature to `String`.

NOTE: This only works for _unconditional_ validations (i.e. ones without `:if` or `:unless`). Here's an example that would still be treated as nilable:
```
#  id          :bigint(8)    not null, primary key
#  first_name  :string(255)
class Student
  validates :first_name, presence: true, unless: :anonymous?
end
```

Because we can't statically be sure that the validation will be called we can't confidently set the type to `String`.

We also do this for:
- belongs_to associations (we check for validations on both the association name OR the foreign key)
- has_one associations (we check for validations on the association name)