Improve array typing (#309)
Add better typing for array types to include the array subtype and also
fix the nullability. Unfortunately there is no easy way to add tests as
the current framework (understandably) uses SQLite for the DB engine
which does not support array types. I tested this with some of our data
models, however, and saw this worked properly:

sample `git diff` for table with `assignee_names` as `:string, array:
true`:
```
-  sig { returns(T::Array[T.nilable(T.untyped)]) }
+  sig { returns(T.nilable(T::Array[String])) }
   def assignee_names; end

-  sig { params(value: T::Array[T.nilable(T.untyped)]).void }
+  sig { params(value: T.nilable(T::Array[T.any(String, Symbol)])).void }
   def assignee_names=(value); end
```