Generate sigs for build/create methods on AssociationRelation and CollectionProxy (#313)
Follow up to #312. The first two commits in this PR are from that one.

`ActiveRecord::AssociationRelation` and `ActiveRecord::Associations::CollectionProxy` both inherit from `ActiveRecord::Relation` so adding sigs for methods on `Relation` should trickle to the others. However, when sorbet autogenerates RBI via reflection is defines an `activerecord.rbi` which defines a few methods on `AssociationRelation` and `CollectionProxy` that take precedence over the inherited ones meaning that we lose sig information.

To get around this, we need to explicitly add some methods to our RBI which is what this PR is doing.

* Uncomment some assertions

* Add sigs for more build/create methods

* Type builder methods in CollectionProxy

* Type builder methods in AssociationRelation

Co-authored-by: Harry Doan <hdoan@chanzuckerberg.com>