Stop infering `validate :on` as unconditional
Recent version of sorbet-rails added a neat optimization to make
attributes non nilable if they had an "unconditional validation".
After updating we had some parts of the code report wrongly unreachable code.
This is becase we have some validations using `:on` which are detected
- wrongly - as unconditional.

Didn't add any test since I saw that only the `:if` clause was actually
tested.