Better handle checking if we can know association type (#322)
I had a model with associations like:
```
has_many :liabilities
has_many :payroll_taxes, through: :liabilities, source: :liability, source_type: 'PayrollTax'
```

This means that payroll_taxes is a "has many through" association where the join table is polymorphic so sorbet-rails was generating:
```ruby
sig { returns(ActiveRecord::Associations::CollectionProxy) }
def payroll_taxes; end
```

I would expect:
```ruby
sig { returns(::PayrollTax::ActiveRecord_Associations_CollectionProxy) }
def payroll_taxes; end
```

This was because it was using logic when checking if the reflection was polymorphic that said "is this a has-many-through? then check if the join table is polymorphic". I don't quite get that logic. In this case it shouldn't matter that the join table is polymorphic (in fact, I actually hadn't realized it was until this came up 😬).

Maybe there's a case you've run into that I'm not considering? If so we should figure out how to handle both ideally 🙂 If not then this PR simplifies the logic a bit.

Co-authored-by: Harry Doan <hdoan@chanzuckerberg.com>