add default url options to url helpers (#338)
Turns out classes including `url_helpers` does not need `default_url_options` defined explicitly to work (See #335). This PR implements a `default_url_options` method that draw it from the Rails application config so that people don't have to implement it explicitly.

Note that the default for `Rails.application.config` is {} and usually applications will set it for their mailer etc. The `_path` methods will work empty options {} but the `_url` methods still require a valid set of options with a HOST.