Don't use #try to get serialization coder
Object#try is defined by ActiveSupport as helper method for calling a method if
it exists. Unfortunately, Dry::Struct#try does something completely different,
so when a model is using that library as a serializer, it fails with the error:

> Error when handling model ExampleModel: can't convert Symbol into Hash

which prevents the rbi file from being generated (and is also extremely
unhelpful as an explanation for the cause).

This change drops the .try call in favor of using the (more verbose but ideally
less naming-conflict-prone) native Ruby method Object#respond_to? to achieve
the same result.